Add files via upload
New and improved comparison.
Seems promising.
Weird issue is that it sometimes prints a line somewhere where is shouldn't be. Might just be because im running it inside a vm.
To be added - maybe a loop that says like if: > 80% and the file has less than 1000 words = this is a good match.